--- src/main/java/com/moyo/backend/follow/domain/DetectType.java
@@ -1,25 +1,0 @@
-package com.moyo.backend.follow.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum DetectType {
-
-    MUTUAL("mutual"),
-    FOLLOW_ONLY("follow-only"),
-    FOLLOWED_ONLY("followed-only")
-    ;
-
-    private final String value;
-
-    public static DetectType fromValue(String value) {
-        for (DetectType type : values()) {
-            if (type.getValue().equalsIgnoreCase(value)) {
-                return type;
-            }
-        }
-        throw new IllegalArgumentException();
-    }
-}
