--- src/main/java/com/moyo/backend/follow/dto/request/GithubFollowDetectRequest.java
@@ -1,14 +1,0 @@
-package com.moyo.backend.follow.dto.request;
-
-import com.moyo.backend.follow.domain.DetectType;
-import lombok.Builder;
-import lombok.Getter;
-@Getter
-@Builder
-public class GithubFollowDetectRequest {
-
-    private DetectType detectType;
-    private Long lastFetchedUserId;
-    private int pagingSize;
-
-}
