+++ src/main/java/com/moyo/backend/githubFollow/domain/DetectType.java
@@ -1,0 +1,25 @@
+package com.moyo.backend.githubFollow.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum DetectType {
+
+    MUTUAL("mutual"),
+    FOLLOW_ONLY("follow-only"),
+    FOLLOWED_ONLY("followed-only")
+    ;
+
+    private final String value;
+
+    public static DetectType fromValue(String value) {
+        for (DetectType type : values()) {
+            if (type.getValue().equalsIgnoreCase(value)) {
+                return type;
+            }
+        }
+        throw new IllegalArgumentException();
+    }
+}
