+++ src/main/java/com/moyo/backend/githubFollow/domain/GithubFollowRelationClient.java
@@ -1,0 +1,10 @@
+package com.moyo.backend.githubFollow.domain;
+
+public interface GithubFollowRelationClient {
+
+    GithubFollowRelation load(Long userId, String username, String accessToken);
+
+    void follow(Long currentUserId, Long targetUserId, String accessToken);
+
+    void unfollow(Long currentUserId, Long targetUserId, String accessToken);
+}
