+++ src/main/java/com/moyo/backend/githubFollow/domain/GithubFollowUser.java
@@ -1,0 +1,9 @@
+package com.moyo.backend.githubFollow.domain;
+
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record GithubFollowUser(@JsonProperty("id") Long id,
+                               @JsonProperty("login") String username,
+                               @JsonProperty("avatar_url") String profileImgUrl) {
+}
